+++ src/components/directory-section/DirectorySection.tsx
@@ -1,5 +1,6 @@
 import styled from 'styled-components';
-import { LightButton } from '../../styles/ButtonStyles';
+import { LightButton } from '../../styles/Button.styles';
+import { ArrowRight } from '../../styles/Icon.styles';
 import { theme } from '../../theme';
 
 interface Props {
@@ -12,6 +13,10 @@ interface Props {
 
 const SectionWrapper = styled.article`
     display: flex;
+    @media (max-width: 480px) {
+        flex-direction: column-reverse;
+        align-items: center;
+    }
 `;
 
 const Image = styled.div<{ photo: string }>`
@@ -20,6 +25,12 @@ const Image = styled.div<{ photo: string }>`
     background-image: url(${(props) => props.photo});
     background-size: cover;
     background-position: center;
+    @media (max-width: 480px) {
+        flex-direction: column-reverse;
+        align-items: center;
+        min-height: 200px;
+        width: 100%;
+    }
 `;
 
 const ContentWrapper = styled.section`
@@ -27,12 +38,30 @@ const ContentWrapper = styled.section`
     flex-direction: column;
     gap: 10px;
     padding: 60px;
-    padding-left: 30px;
+    padding-left: ${theme.spacing.large};
     text-align: left;
     width: 30%;
     background: ${theme.colors.secondary};
     color: ${theme.colors.white};
     border-radius: 0 ${theme.borderRadius.small} ${theme.borderRadius.small} 0;
+    @media (max-width: 480px) {
+        width: 100%;
+        padding: ${theme.spacing.small};
+        flex-direction: row;
+        align-items: center;
+    }
+`;
+
+const Button = styled(LightButton)`
+    @media (max-width: 480px) {
+        display: none;
+    }
+`;
+
+const Description = styled.article`
+    @media (max-width: 480px) {
+        display: none;
+    }
 `;
 
 const Title = styled.h2`
@@ -51,8 +80,9 @@ function DirectorySection({
             <Image photo={photo} />
             <ContentWrapper>
                 <Title>{title}</Title>
-                <p>{description}</p>
-                <LightButton onClick={onClick}>{actionLabel}</LightButton>
+                <Description>{description}</Description>
+                <Button onClick={onClick}>{actionLabel}</Button>
+                <ArrowRight />
             </ContentWrapper>
         </SectionWrapper>
     );
