+++ src/components/search-input/SearchInput.tsx
@@ -1,5 +1,5 @@
 import React, { useCallback, useState } from 'react';
-import { OptionItem, OptionsList } from '../../styles/listStyles';
+import { OptionItem, OptionsList } from '../../styles/list.styles';
 import { simpleFuzzySearch } from '../../utils/fuzzy-search';
 import {
     Input,
