+++ src/styles/Global.styles.ts
