+++ src/styles/list.styles.ts
